refactor(sns): split service.rs into resource-grouped impl files
service.rs grew to 3152 lines after Phase I/J test extraction. Move
publish/deliver_to_*, platform_application/endpoint, and SMS sandbox
methods into sibling files (service_publish.rs, service_platform.rs,
service_sms.rs) loaded via #[path] modules with impl SnsService blocks.
Methods exposed as pub(super) so the test module can still reach them.

service.rs: 3152 -> 1564 lines (-50%).
service_publish.rs: 783, service_platform.rs: 462, service_sms.rs: 419.

Closes audit-2026-04-28 §1 carryover for sns/service.rs >2000 lines.
