+++ crates/fakecloud-sns/src/service_publish.rs
@@ -1,0 +1,788 @@
+// Auto-extracted from service.rs as part of carryover service.rs split.
+// Methods on `SnsService` grouped by resource concern.
+
+#![allow(clippy::too_many_arguments)]
+
+use base64::Engine;
+use chrono::Utc;
+use http::StatusCode;
+use serde_json::Value;
+use std::collections::{BTreeMap, HashMap};
+
+use fakecloud_core::service::{AwsRequest, AwsResponse, AwsServiceError};
+
+use super::*;
+
+impl SnsService {
+    pub(super) fn publish(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
+        // Either TopicArn or TargetArn is required; also allow PhoneNumber for SMS
+        let topic_arn = param(req, "TopicArn").or_else(|| param(req, "TargetArn"));
+        let phone_number = param(req, "PhoneNumber");
+
+        if topic_arn.is_none() && phone_number.is_none() {
+            return Err(AwsServiceError::aws_error(
+                StatusCode::BAD_REQUEST,
+                "InvalidParameter",
+                "The request must contain the parameter TopicArn or TargetArn or PhoneNumber",
+            ));
+        }
+
+        let message = required(req, "Message")?;
+        let subject = param(req, "Subject");
+        let message_group_id = param(req, "MessageGroupId");
+        let message_dedup_id = param(req, "MessageDeduplicationId");
+        let message_structure = param(req, "MessageStructure");
+
+        // Validate subject length
+        if let Some(ref subj) = subject {
+            if subj.len() > 100 {
+                return Err(AwsServiceError::aws_error(
+                    StatusCode::BAD_REQUEST,
+                    "InvalidParameter",
+                    "Subject must be less than 100 characters",
+                ));
+            }
+        }
+
+        // Validate message length (256KB)
+        if message.len() > 262144 {
+            return Err(AwsServiceError::aws_error(
+                StatusCode::BAD_REQUEST,
+                "InvalidParameter",
+                "Invalid parameter: Message too long",
+            ));
+        }
+
+        // Validate MessageStructure=json
+        if message_structure.as_deref() == Some("json") {
+            validate_message_structure_json(&message)?;
+        }
+
+        // Parse MessageAttributes from query params
+        let message_attributes = parse_message_attributes(req);
+
+        if let Some(ref phone) = phone_number {
+            return self.publish_to_phone_number(
+                req,
+                phone,
+                message,
+                subject,
+                message_attributes,
+                message_group_id,
+                message_dedup_id,
+            );
+        }
+
+        let topic_arn = topic_arn.ok_or_else(|| {
+            AwsServiceError::aws_error(
+                StatusCode::BAD_REQUEST,
+                "InvalidParameter",
+                "TopicArn or TargetArn is required",
+            )
+        })?;
+
+        // Check if it's a platform endpoint ARN
+        if topic_arn.contains(":endpoint/") {
+            return self.publish_to_platform_endpoint(
+                &topic_arn,
+                &message,
+                &message_attributes,
+                &req.request_id,
+            );
+        }
+
+        let mut accounts = self.state.write();
+        let state = accounts.get_or_create(&req.account_id);
+        let topic = state
+            .topics
+            .get(&topic_arn)
+            .ok_or_else(|| not_found("Topic"))?;
+
+        // FIFO topic enforcement
+        if topic.is_fifo {
+            if message_group_id.is_none() {
+                return Err(AwsServiceError::aws_error(
+                    StatusCode::BAD_REQUEST,
+                    "InvalidParameter",
+                    "Invalid parameter: The request must contain the parameter MessageGroupId.",
+                ));
+            }
+            // FIFO topics require deduplication: either ContentBasedDeduplication or explicit ID
+            let content_dedup = topic
+                .attributes
+                .get("ContentBasedDeduplication")
+                .map(|v| v == "true")
+                .unwrap_or(false);
+            if !content_dedup && message_dedup_id.is_none() {
+                return Err(AwsServiceError::aws_error(
+                    StatusCode::BAD_REQUEST,
+                    "InvalidParameter",
+                    "Invalid parameter: The topic should either have ContentBasedDeduplication enabled or MessageDeduplicationId provided explicitly",
+                ));
+            }
+        } else {
+            // Non-FIFO: MessageGroupId is allowed (forwarded to SQS for fair queuing)
+            // But DeduplicationId is NOT allowed on non-FIFO topics
+            if message_dedup_id.is_some() {
+                return Err(AwsServiceError::aws_error(
+                    StatusCode::BAD_REQUEST,
+                    "InvalidParameter",
+                    "Invalid parameter: The request includes MessageDeduplicationId parameter that is not valid for this topic type",
+                ));
+            }
+        }
+
+        let kms_key_id = topic
+            .attributes
+            .get("KmsMasterKeyId")
+            .cloned()
+            .filter(|s| !s.is_empty());
+        self.record_topic_kms_usage(&req.account_id, &topic_arn, kms_key_id.as_deref(), &message)?;
+
+        let msg_id = uuid::Uuid::new_v4().to_string();
+        state.published.push(PublishedMessage {
+            message_id: msg_id.clone(),
+            topic_arn: topic_arn.clone(),
+            message: message.clone(),
+            subject: subject.clone(),
+            message_attributes: message_attributes.clone(),
+            message_group_id: message_group_id.clone(),
+            message_dedup_id: message_dedup_id.clone(),
+            timestamp: Utc::now(),
+        });
+
+        // Resolve the actual message per protocol for MessageStructure=json
+        let parsed_structure: Option<Value> = if message_structure.as_deref() == Some("json") {
+            Some(serde_json::from_str(&message).map_err(|_| {
+                AwsServiceError::aws_error(
+                    StatusCode::BAD_REQUEST,
+                    "InvalidParameter",
+                    "Invalid parameter: Message Structure - No JSON message body is parseable",
+                )
+            })?)
+        } else {
+            None
+        };
+
+        let subscribers =
+            collect_topic_subscribers(state, &topic_arn, &message_attributes, &message);
+        let endpoint = state.endpoint.clone();
+        drop(accounts);
+
+        // Determine actual message content per protocol
+        let sqs_message = if let Some(ref structure) = parsed_structure {
+            structure
+                .get("sqs")
+                .or_else(|| structure.get("default"))
+                .and_then(|v| v.as_str())
+                .unwrap_or(&message)
+                .to_string()
+        } else {
+            message.clone()
+        };
+
+        let default_message = if let Some(ref structure) = parsed_structure {
+            structure
+                .get("default")
+                .and_then(|v| v.as_str())
+                .unwrap_or(&message)
+                .to_string()
+        } else {
+            message.clone()
+        };
+
+        let envelope_attrs = build_envelope_attrs(&message_attributes);
+
+        let ctx = TopicFanoutContext {
+            msg_id: &msg_id,
+            topic_arn: &topic_arn,
+            subject: subject.as_deref(),
+            endpoint: &endpoint,
+            sqs_message: &sqs_message,
+            default_message: &default_message,
+            envelope_attrs: &envelope_attrs,
+            message_attributes: &message_attributes,
+            message_group_id: message_group_id.as_deref(),
+            message_dedup_id: message_dedup_id.as_deref(),
+        };
+
+        self.deliver_to_sqs_subscribers(&subscribers.sqs, &ctx);
+        self.deliver_to_http_subscribers(&subscribers.http, &ctx);
+        self.deliver_to_lambda_subscribers(&subscribers.lambda, &ctx);
+        self.deliver_to_email_subscribers(&subscribers.email, &ctx);
+        self.deliver_to_sms_subscribers(&subscribers.sms, &ctx);
+
+        Ok(xml_resp(
+            &format!(
+                r#"<PublishResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/">
+  <PublishResult>
+    <MessageId>{msg_id}</MessageId>
+  </PublishResult>
+  <ResponseMetadata>
+    <RequestId>{}</RequestId>
+  </ResponseMetadata>
+</PublishResponse>"#,
+                req.request_id
+            ),
+            &req.request_id,
+        ))
+    }
+
+    pub(super) fn deliver_to_sqs_subscribers(
+        &self,
+        subs: &[(String, bool)],
+        ctx: &TopicFanoutContext<'_>,
+    ) {
+        for (queue_arn, raw) in subs {
+            if *raw {
+                let mut sqs_msg_attrs = HashMap::new();
+                for (k, v) in ctx.message_attributes {
+                    let mut attr = fakecloud_core::delivery::SqsMessageAttribute {
+                        data_type: v.data_type.clone(),
+                        string_value: v.string_value.clone(),
+                        binary_value: None,
+                    };
+                    if let Some(ref bv) = v.binary_value {
+                        attr.binary_value =
+                            Some(base64::engine::general_purpose::STANDARD.encode(bv));
+                    }
+                    sqs_msg_attrs.insert(k.clone(), attr);
+                }
+                self.delivery.send_to_sqs_with_attrs(
+                    queue_arn,
+                    ctx.sqs_message,
+                    &sqs_msg_attrs,
+                    ctx.message_group_id,
+                    ctx.message_dedup_id,
+                );
+            } else {
+                let envelope_str = build_sns_envelope(
+                    ctx.msg_id,
+                    ctx.topic_arn,
+                    &ctx.subject.map(|s| s.to_string()),
+                    ctx.sqs_message,
+                    ctx.envelope_attrs,
+                    ctx.endpoint,
+                );
+                self.delivery
+                    .send_to_sqs(queue_arn, &envelope_str, &HashMap::new());
+            }
+        }
+    }
+
+    pub(super) fn deliver_to_http_subscribers(
+        &self,
+        subs: &[String],
+        ctx: &TopicFanoutContext<'_>,
+    ) {
+        for endpoint_url in subs {
+            let body = build_sns_envelope(
+                ctx.msg_id,
+                ctx.topic_arn,
+                &ctx.subject.map(|s| s.to_string()),
+                ctx.default_message,
+                ctx.envelope_attrs,
+                ctx.endpoint,
+            );
+            let endpoint_url = endpoint_url.clone();
+            let topic = ctx.topic_arn.to_string();
+            tokio::spawn(async move {
+                let client = reqwest::Client::new();
+                let result = client
+                    .post(&endpoint_url)
+                    .header("Content-Type", "application/json")
+                    .header("x-amz-sns-message-type", "Notification")
+                    .header("x-amz-sns-topic-arn", &topic)
+                    .body(body)
+                    .send()
+                    .await;
+                if let Err(e) = result {
+                    tracing::warn!(endpoint = %endpoint_url, error = %e, "SNS HTTP delivery failed");
+                }
+            });
+        }
+    }
+
+    pub(super) fn deliver_to_lambda_subscribers(
+        &self,
+        subs: &[(String, String)],
+        ctx: &TopicFanoutContext<'_>,
+    ) {
+        if subs.is_empty() {
+            return;
+        }
+        let now = Utc::now();
+        let subject_owned = ctx.subject.map(|s| s.to_string());
+
+        let lambda_payloads: Vec<(String, String)> = subs
+            .iter()
+            .map(|(function_arn, subscription_arn)| {
+                let payload = build_sns_lambda_event(&SnsLambdaEventInput {
+                    message_id: ctx.msg_id,
+                    topic_arn: ctx.topic_arn,
+                    subscription_arn,
+                    message: ctx.default_message,
+                    subject: ctx.subject,
+                    message_attributes: ctx.envelope_attrs,
+                    timestamp: &now,
+                    endpoint: ctx.endpoint,
+                });
+                (function_arn.clone(), payload)
+            })
+            .collect();
+
+        {
+            let acct = ctx.topic_arn.split(':').nth(4).unwrap_or("");
+            let mut accounts = self.state.write();
+            let state = accounts.get_or_create(acct);
+            for (function_arn, _) in &lambda_payloads {
+                state
+                    .lambda_invocations
+                    .push(crate::state::LambdaInvocation {
+                        function_arn: function_arn.clone(),
+                        message: ctx.default_message.to_string(),
+                        subject: subject_owned.clone(),
+                        timestamp: now,
+                    });
+            }
+        }
+
+        let delivery = self.delivery.clone();
+        tokio::spawn(async move {
+            for (function_arn, payload) in lambda_payloads {
+                tracing::info!(function_arn = %function_arn, "SNS invoking Lambda function");
+                match delivery.invoke_lambda(&function_arn, &payload).await {
+                    Some(Ok(_)) => {
+                        tracing::info!(
+                            function_arn = %function_arn,
+                            "SNS->Lambda invocation succeeded"
+                        );
+                    }
+                    Some(Err(e)) => {
+                        tracing::error!(
+                            function_arn = %function_arn,
+                            error = %e,
+                            "SNS->Lambda invocation failed"
+                        );
+                    }
+                    None => {
+                        tracing::debug!(
+                            function_arn = %function_arn,
+                            "SNS->Lambda: no container runtime, skipping real execution"
+                        );
+                    }
+                }
+            }
+        });
+    }
+
+    pub(super) fn deliver_to_email_subscribers(
+        &self,
+        subs: &[String],
+        ctx: &TopicFanoutContext<'_>,
+    ) {
+        if subs.is_empty() {
+            return;
+        }
+        let now = Utc::now();
+        let subject_owned = ctx.subject.map(|s| s.to_string());
+        let acct = ctx.topic_arn.split(':').nth(4).unwrap_or("");
+        let mut accounts = self.state.write();
+        let state = accounts.get_or_create(acct);
+        for email_address in subs {
+            tracing::info!(
+                email = %email_address,
+                topic_arn = %ctx.topic_arn,
+                "SNS delivering to email (stub)"
+            );
+            state.sent_emails.push(crate::state::SentEmail {
+                email_address: email_address.clone(),
+                message: ctx.default_message.to_string(),
+                subject: subject_owned.clone(),
+                topic_arn: ctx.topic_arn.to_string(),
+                timestamp: now,
+            });
+        }
+    }
+
+    pub(super) fn deliver_to_sms_subscribers(&self, subs: &[String], ctx: &TopicFanoutContext<'_>) {
+        if subs.is_empty() {
+            return;
+        }
+        let acct = ctx.topic_arn.split(':').nth(4).unwrap_or("");
+        let mut accounts = self.state.write();
+        let state = accounts.get_or_create(acct);
+        for phone_number in subs {
+            tracing::info!(
+                phone_number = %phone_number,
+                topic_arn = %ctx.topic_arn,
+                "SNS delivering to SMS (stub)"
+            );
+            state
+                .sms_messages
+                .push((phone_number.clone(), ctx.default_message.to_string()));
+        }
+    }
+
+    pub(super) fn publish_batch(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
+        let topic_arn = required(req, "TopicArn")?;
+
+        let _accts = self.state.read();
+        let _empty = crate::state::SnsState::new(&req.account_id, &req.region, "");
+        let state = _accts.get(&req.account_id).unwrap_or(&_empty);
+        let topic = state
+            .topics
+            .get(&topic_arn)
+            .ok_or_else(|| not_found("Topic"))?;
+        let is_fifo = topic.is_fifo;
+        let endpoint = state.endpoint.clone();
+        drop(_accts);
+
+        // Parse batch entries: PublishBatchRequestEntries.member.N.*
+        let mut entries = Vec::new();
+        for n in 1..=100 {
+            let id_key = format!("PublishBatchRequestEntries.member.{n}.Id");
+            if let Some(id) = req.query_params.get(&id_key) {
+                let msg_key = format!("PublishBatchRequestEntries.member.{n}.Message");
+                let message = req.query_params.get(&msg_key).cloned().unwrap_or_default();
+                let subject_key = format!("PublishBatchRequestEntries.member.{n}.Subject");
+                let subject = req.query_params.get(&subject_key).cloned();
+                let group_key = format!("PublishBatchRequestEntries.member.{n}.MessageGroupId");
+                let group_id = req.query_params.get(&group_key).cloned();
+                let dedup_key =
+                    format!("PublishBatchRequestEntries.member.{n}.MessageDeduplicationId");
+                let dedup_id = req.query_params.get(&dedup_key).cloned();
+                let structure_key =
+                    format!("PublishBatchRequestEntries.member.{n}.MessageStructure");
+                let message_structure = req.query_params.get(&structure_key).cloned();
+                entries.push((
+                    id.clone(),
+                    message,
+                    subject,
+                    group_id,
+                    dedup_id,
+                    message_structure,
+                ));
+            } else {
+                break;
+            }
+        }
+
+        // Validate: max 10 entries
+        if entries.len() > 10 {
+            return Err(AwsServiceError::aws_error(
+                StatusCode::BAD_REQUEST,
+                "TooManyEntriesInBatchRequest",
+                "The batch request contains more entries than permissible.",
+            ));
+        }
+
+        // Validate: unique IDs
+        let ids: Vec<&str> = entries.iter().map(|e| e.0.as_str()).collect();
+        let unique_ids: std::collections::HashSet<&str> = ids.iter().copied().collect();
+        if unique_ids.len() != ids.len() {
+            return Err(AwsServiceError::aws_error(
+                StatusCode::BAD_REQUEST,
+                "BatchEntryIdsNotDistinct",
+                "Two or more batch entries in the request have the same Id.",
+            ));
+        }
+
+        // FIFO: all entries must have MessageGroupId — this is a top-level error
+        if is_fifo && entries.iter().any(|e| e.3.is_none()) {
+            return Err(AwsServiceError::aws_error(
+                StatusCode::BAD_REQUEST,
+                "InvalidParameter",
+                "Invalid parameter: The MessageGroupId parameter is required for FIFO topics",
+            ));
+        }
+
+        let mut successful = Vec::new();
+        let failed: Vec<String> = Vec::new();
+
+        for (idx, (id, message, subject, group_id, dedup_id, structure)) in
+            entries.iter().enumerate()
+        {
+            // Parse per-entry message attributes
+            let batch_attrs = parse_batch_message_attributes(req, idx + 1);
+
+            // Validate MessageStructure=json
+            if structure.as_deref() == Some("json") {
+                validate_message_structure_json(message)?;
+            }
+
+            let msg_id = uuid::Uuid::new_v4().to_string();
+            let mut accounts = self.state.write();
+            let state = accounts.get_or_create(&req.account_id);
+            state.published.push(PublishedMessage {
+                message_id: msg_id.clone(),
+                topic_arn: topic_arn.clone(),
+                message: message.clone(),
+                subject: subject.clone(),
+                message_attributes: batch_attrs.clone(),
+                message_group_id: group_id.clone(),
+                message_dedup_id: dedup_id.clone(),
+                timestamp: Utc::now(),
+            });
+
+            // Resolve message for SQS via MessageStructure=json
+            let parsed_structure: Option<Value> = if structure.as_deref() == Some("json") {
+                Some(serde_json::from_str(message).map_err(|_| {
+                    AwsServiceError::aws_error(
+                        StatusCode::BAD_REQUEST,
+                        "InvalidParameter",
+                        "Invalid parameter: Message Structure - No JSON message body is parseable",
+                    )
+                })?)
+            } else {
+                None
+            };
+            let sqs_message = if let Some(ref s) = parsed_structure {
+                s.get("sqs")
+                    .or_else(|| s.get("default"))
+                    .and_then(|v| v.as_str())
+                    .unwrap_or(message)
+                    .to_string()
+            } else {
+                message.clone()
+            };
+
+            // Deliver to SQS subscribers
+            let sqs_subscribers: Vec<(String, bool)> = state
+                .subscriptions
+                .values()
+                .filter(|s| s.topic_arn == topic_arn && s.protocol == "sqs" && s.confirmed)
+                .map(|s| {
+                    let raw = s
+                        .attributes
+                        .get("RawMessageDelivery")
+                        .map(|v| v == "true")
+                        .unwrap_or(false);
+                    (s.endpoint.clone(), raw)
+                })
+                .collect();
+            drop(accounts);
+
+            // Build envelope attributes
+            let mut envelope_attrs = serde_json::Map::new();
+            for (key, attr) in &batch_attrs {
+                let mut attr_obj = serde_json::Map::new();
+                attr_obj.insert("Type".to_string(), Value::String(attr.data_type.clone()));
+                if let Some(ref sv) = attr.string_value {
+                    attr_obj.insert("Value".to_string(), Value::String(sv.clone()));
+                }
+                if let Some(ref bv) = attr.binary_value {
+                    attr_obj.insert(
+                        "Value".to_string(),
+                        Value::String(base64::engine::general_purpose::STANDARD.encode(bv)),
+                    );
+                }
+                envelope_attrs.insert(key.clone(), Value::Object(attr_obj));
+            }
+
+            for (queue_arn, raw) in &sqs_subscribers {
+                if *raw {
+                    let mut sqs_msg_attrs = HashMap::new();
+                    for (k, v) in &batch_attrs {
+                        let mut attr = fakecloud_core::delivery::SqsMessageAttribute {
+                            data_type: v.data_type.clone(),
+                            string_value: v.string_value.clone(),
+                            binary_value: None,
+                        };
+                        if let Some(ref bv) = v.binary_value {
+                            attr.binary_value =
+                                Some(base64::engine::general_purpose::STANDARD.encode(bv));
+                        }
+                        sqs_msg_attrs.insert(k.clone(), attr);
+                    }
+                    self.delivery.send_to_sqs_with_attrs(
+                        queue_arn,
+                        &sqs_message,
+                        &sqs_msg_attrs,
+                        if is_fifo { group_id.as_deref() } else { None },
+                        if is_fifo { dedup_id.as_deref() } else { None },
+                    );
+                } else {
+                    let envelope_str = build_sns_envelope(
+                        &msg_id,
+                        &topic_arn,
+                        subject,
+                        &sqs_message,
+                        &envelope_attrs,
+                        &endpoint,
+                    );
+                    self.delivery
+                        .send_to_sqs(queue_arn, &envelope_str, &HashMap::new());
+                }
+            }
+
+            successful.push(format!(
+                r#"    <member>
+      <Id>{id}</Id>
+      <MessageId>{msg_id}</MessageId>
+    </member>"#
+            ));
+        }
+
+        let successful_xml = successful.join("\n");
+        let failed_xml = failed.join("\n");
+
+        Ok(xml_resp(
+            &format!(
+                r#"<PublishBatchResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/">
+  <PublishBatchResult>
+    <Successful>
+{successful_xml}
+    </Successful>
+    <Failed>
+{failed_xml}
+    </Failed>
+  </PublishBatchResult>
+  <ResponseMetadata>
+    <RequestId>{}</RequestId>
+  </ResponseMetadata>
+</PublishBatchResponse>"#,
+                req.request_id
+            ),
+            &req.request_id,
+        ))
+    }
+
+    /// Publish directly to an SMS destination (`Publish` called with
+    /// `PhoneNumber` instead of `TopicArn` / `TargetArn`). This path
+    /// does its own length and E.164 validation since AWS reports
+    /// distinct error messages for SMS.
+    #[allow(clippy::too_many_arguments)]
+    pub(super) fn publish_to_phone_number(
+        &self,
+        req: &AwsRequest,
+        phone: &str,
+        message: String,
+        subject: Option<String>,
+        message_attributes: BTreeMap<String, MessageAttribute>,
+        message_group_id: Option<String>,
+        message_dedup_id: Option<String>,
+    ) -> Result<AwsResponse, AwsServiceError> {
+        let is_valid_e164 = phone.starts_with('+')
+            && phone.len() >= 2
+            && phone[1..].chars().all(|c| c.is_ascii_digit());
+        if !is_valid_e164 {
+            return Err(AwsServiceError::aws_error(
+                StatusCode::BAD_REQUEST,
+                "InvalidParameter",
+                format!(
+                    "Invalid parameter: PhoneNumber Reason: {phone} does not meet the E164 format"
+                ),
+            ));
+        }
+
+        if message.len() > 1600 {
+            return Err(AwsServiceError::aws_error(
+                StatusCode::BAD_REQUEST,
+                "InvalidParameter",
+                "Invalid parameter: Message Reason: Message must be less than 1600 characters long",
+            ));
+        }
+
+        let msg_id = uuid::Uuid::new_v4().to_string();
+        let mut accounts = self.state.write();
+        let state = accounts.get_or_create(&req.account_id);
+        state
+            .sms_messages
+            .push((phone.to_string(), message.clone()));
+        state.published.push(PublishedMessage {
+            message_id: msg_id.clone(),
+            topic_arn: String::new(),
+            message,
+            subject,
+            message_attributes,
+            message_group_id,
+            message_dedup_id,
+            timestamp: Utc::now(),
+        });
+
+        Ok(xml_resp(
+            &format!(
+                r#"<PublishResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/">
+  <PublishResult>
+    <MessageId>{msg_id}</MessageId>
+  </PublishResult>
+  <ResponseMetadata>
+    <RequestId>{}</RequestId>
+  </ResponseMetadata>
+</PublishResponse>"#,
+                req.request_id
+            ),
+            &req.request_id,
+        ))
+    }
+
+    pub(super) fn publish_to_platform_endpoint(
+        &self,
+        endpoint_arn: &str,
+        message: &str,
+        message_attributes: &BTreeMap<String, MessageAttribute>,
+        request_id: &str,
+    ) -> Result<AwsResponse, AwsServiceError> {
+        let acct = endpoint_arn.split(':').nth(4).unwrap_or("");
+        let _accts = self.state.read();
+        let state = _accts.get(acct).unwrap_or_else(|| _accts.default_ref());
+
+        // Find the platform endpoint
+        let mut found_endpoint: Option<&PlatformEndpoint> = None;
+        for app in state.platform_applications.values() {
+            if let Some(ep) = app.endpoints.get(endpoint_arn) {
+                found_endpoint = Some(ep);
+                break;
+            }
+        }
+
+        let ep = found_endpoint.ok_or_else(|| {
+            AwsServiceError::aws_error(StatusCode::NOT_FOUND, "NotFound", "Endpoint does not exist")
+        })?;
+
+        if !ep.enabled {
+            return Err(AwsServiceError::aws_error(
+                StatusCode::BAD_REQUEST,
+                "EndpointDisabled",
+                "Endpoint is disabled",
+            ));
+        }
+        drop(_accts);
+
+        let msg_id = uuid::Uuid::new_v4().to_string();
+        let acct = endpoint_arn.split(':').nth(4).unwrap_or("");
+        let mut accounts = self.state.write();
+        let state = accounts.get_or_create(acct);
+        // Store message on the endpoint
+        for app in state.platform_applications.values_mut() {
+            if let Some(ep) = app.endpoints.get_mut(endpoint_arn) {
+                ep.messages.push(PublishedMessage {
+                    message_id: msg_id.clone(),
+                    topic_arn: endpoint_arn.to_string(),
+                    message: message.to_string(),
+                    subject: None,
+                    message_attributes: message_attributes.clone(),
+                    message_group_id: None,
+                    message_dedup_id: None,
+                    timestamp: Utc::now(),
+                });
+                break;
+            }
+        }
+
+        Ok(xml_resp(
+            &format!(
+                r#"<PublishResponse xmlns="http://sns.amazonaws.com/doc/2010-03-31/">
+  <PublishResult>
+    <MessageId>{msg_id}</MessageId>
+  </PublishResult>
+  <ResponseMetadata>
+    <RequestId>{request_id}</RequestId>
+  </ResponseMetadata>
+</PublishResponse>"#,
+            ),
+            request_id,
+        ))
+    }
+}
